Fixed a typo in the HTML Element, where the input ids were mistakenly typed as "is".
Also few minor bug fixes and refactoring.

ADDITIONAL NOTES:
=================

Issues:
When an alarm is added in the list, it should display the data in time format (HH:MM:SS),
but all it displays is, "undefined undefined undefined". Maybe for some reason it
doesn't extract the hour, minute, and second, like it should. What should I do?
(The code block for that, is from line 40 to 60)